+++ R/tryCatchLog.R
@@ -33,7 +33,9 @@
 #'                              All condition handlers are passed to \code{\link{tryCatch}} as is
 #'                              (no filtering, wrapping or changing of semantics).
 #' @param finally               expression to be evaluated at the end
-#' @param write.error.dump.file \code{TRUE}: Saves a dump of the workspace and the call stack named \code{dump_<YYYYMMDD_HHMMSS>.rda}
+#' @param write.error.dump.file \code{TRUE}: Saves a dump of the workspace and the call stack named
+#'                              \code{dump_<YYYYMMDD>_at_<HHMMSS.sss>_PID_<process id>.rda}.
+#'                              This dump file name pattern shall ensure unique file names in parallel processing scenarios.
 #' @param write.error.dump.folder    \code{path}: Saves the dump of the workspace in a specific folder instead of the working directory
 #' @param silent.warnings       \code{TRUE}: Warnings are logged, but not propagated to the caller.\cr
 #'                              \code{FALSE}: Warnings are logged and treated according to the global
@@ -171,11 +173,17 @@ tryCatchLog <- function(expr,
       # https://bugs.r-project.org/bugzilla/show_bug.cgi?id=17116
       # An enhanced version of "dump.frames" was released in spring 2017 but does still not fulfill the requirements of tryCatchLog:
       # dump.frames(dumpto = dump.file.name, to.file = TRUE, include.GlobalEnv = TRUE)  # test it yourself!
-      dump.file.name  <- format(timestamp, format = "dump_%Y%m%d_%H%M%S.rda")   # use %OS3 (= seconds incl. milliseconds) for finer precision
-      # TODO assert valid folder name
+      # See ?strptime for the available formatting codes...
+      #
+      # Creates a (hopefully) unique dump file name even in case of multiple parallel processes
+      # or multiple sequential errors in the same R process.
+      # Fixes issue #39 by appending fractional seconds (milliseconds) and the process id (PID)
+      # https://github.com/aryoda/tryCatchLog/issues/39
+      # Example dump file name: dump_2019-03-13_at_15-39-33.086_PID_15270.rda
+      dump.file.name  <- paste0(format(timestamp, format = "dump_%Y-%m-%d_at_%H-%M-%OS3"), "_PID_", Sys.getpid(), ".rda")  # %OS3 (= seconds incl. milliseconds)
       dir.create(path <- write.error.dump.folder, recursive = T, showWarnings = F)
       utils::dump.frames()
-      save.image(file = file.path(write.error.dump.folder, dump.file.name))
+      save.image(file = file.path(write.error.dump.folder, dump.file.name))  # an existing file would be overwritten silently :-()
     }